:art: FLOW-647 group management code segregated
